Move all state/tx generation into Hydra.Direct.Chain.State
Only this way we can avoid orphan instances and cyclic module dependencies.
